+++ fissure/Server/TargetSignalIdentification.py
@@ -47,7 +47,7 @@ class TargetSignalIdentification:
 
     settings: Dict
     identifier: str = fissure.comms.Identifiers.TSI
-    logger: logging.Logger = fissure.utils.get_logger(fissure.comms.Identifiers.TSI)
+    # logger: logging.Logger = fissure.utils.get_logger(fissure.comms.Identifiers.TSI)
     ip_address: str
     hiprfisr_address: str
     hiprfisr_socket: fissure.comms.Listener
@@ -59,9 +59,18 @@ class TargetSignalIdentification:
 
 
     def __init__(self):
+        self.logger = fissure.utils.get_logger(fissure.comms.Identifiers.TSI)
+        self.logger.info("=== INITIALIZING ===")
 
-        self.logger.debug("=== INITIALIZING ===")
         self.settings = fissure.utils.get_fissure_config()
+
+        # Update Logging Levels
+        fissure.utils.update_logging_levels(
+            self.logger, 
+            self.settings["console_logging_level"], 
+            self.settings["file_logging_level"]
+        )
+
         self.ip_address = "localhost"
         self.os_info = fissure.utils.get_os_info()
 
@@ -244,26 +253,8 @@ class TargetSignalIdentification:
         """
         Update the logging levels on TSI.
         """
-        # Update New Levels for the HIPRFISR
-        for n in range(0, len(self.logger.parent.handlers)):
-            if self.logger.parent.handlers[n].name == "console":
-                if new_console_level == "DEBUG":
-                    self.logger.parent.handlers[n].level = 10
-                elif new_console_level == "INFO":
-                    self.logger.parent.handlers[n].level = 20
-                elif new_console_level == "WARNING":
-                    self.logger.parent.handlers[n].level = 30
-                elif new_console_level == "ERROR":
-                    self.logger.parent.handlers[n].level = 40
-            elif self.logger.parent.handlers[n].name == "file":
-                if new_file_level == "DEBUG":
-                    self.logger.parent.handlers[n].level = 10
-                elif new_file_level == "INFO":
-                    self.logger.parent.handlers[n].level = 20
-                elif new_file_level == "WARNING":
-                    self.logger.parent.handlers[n].level = 30
-                elif new_file_level == "ERROR":
-                    self.logger.parent.handlers[n].level = 40
+        # Update New Levels for TSI
+        fissure.utils.update_logging_levels(self.logger, new_console_level, new_file_level)
 
 
     def startTSI_ConditionerThread(